feat: move preload to arkhi package
