+++ pages/test/index.page.tsx
@@ -1,8 +1,6 @@
 import React from "react";
 import "./code.css";
 import { Link } from "#/arkhi/client";
-import Games from "../../content/games.mdx";
-import Index from "../../content/index.md";
 import MdxContentRenderer from "./MdxContentRenderer";
 import "./code.css";
 export { Page };
